start stack chart: create data elements
A Stack Chart has a three-tier model: a StackEntry contains zero or
more StackUnit's, which contain zero or more StackItems.

Each stack element has a minimum and a maximum value. It is assumed
that the min and max bounds of a child are within the bounds of its
parent, e.g. stackEntry.max >= stackUnit.max >= stackItem.max.

Add extensions for current and future commits.